fixed sandbox with new db
