+++ mock/jsons/bookings.json
@@ -1,7 +1,7 @@
 [
   {
+    "stockIndex": 0,
     "token": "2ALYY4",
-    "userEmail": "pctest.jeune.93@btmx.fr",
-    "offerIndex": 0
+    "userEmail": "pctest.jeune.93@btmx.fr"
   }
 ]
