+++ mock/jsons/event_occurrences.json
@@ -1,6 +1,5 @@
 [
   {
-    "eventIndex": 0,
-    "offererIndex": 0
+    "offerIndex": 0
   }
 ]
