+++ mock/jsons/offers.json
@@ -1,0 +1,6 @@
+[
+  {
+    "eventIndex": 0,
+    "venueIndex": 0
+  }
+]
